Remove field `documentId` from structures and store this data in field `id` instead.
Previously, somewhere in code we had stored simultaneously `id` and `documentId`. This is useless, we need only one of them.
